+++ NEWS.md
@@ -7,6 +7,10 @@
   - Updates to factor and logical filters now match the initial render (@mikmart #975).
   - Disabled status is now also affected by updates (@mikmart #977).
 
+## BUG FIXES
+
+- Fix a bug that the column filter didn't work for strings like "+" (@stephan-hutter @shrektan #980).
+
 # CHANGES IN DT VERSION 0.21
 
 ## NEW FEATURES