+++ NEWS.md
@@ -1,3 +1,6 @@
+# CHANGES IN DT VERSION 0.23
+
+
 # CHANGES IN DT VERSION 0.22
 
 - Enhancements to `updateFilters()` (#971):